Changed : Changed sub logic

+++ quality-gate-test-app/src/main/java/com/test/service/CalculatorService.java
@@ -9,7 +9,7 @@ public class CalculatorService {
         return a + b;
     }
 
-    public double subtract(double a, double b) {
+    public int subtract(int a, int b) {
         return a - b;
     }
